pyproject.toml
[tool.poetry]
name = "OrdoAbChaos"
version = "0.1.0"
description = ""
authors = ["Yartsev Vladimir <v.rainovsky@yandex.ru>"]
readme = "README.md"
packages = [
    { include = "settings" },
    { include = "sql" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras = ["standart"], version = "^0.21.1"}
strawberry-graphql = {extras = ["debug-server"], version = "^0.168.1"}
sqlalchemy = "^2.0.7"
asyncmy = "^0.2.7"
loguru = "^0.6.0"
isoweek = "^1.3.3"

[tool.poetry.scripts]
recreate_tables = "scripts:recreate_tables"
run_server = "server:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
